Enhance unique ID generation and improve item removal logic across views. Updated getUniqueId function to handle negative IDs, added new row functionality in EmployeeView, and refined removeItem methods in ApplicationsView and PortfolioView to filter out items with negative IDs.

+++ src/views/PortfolioView.vue
@@ -50,7 +50,12 @@ const updateRow = async (row: IPortfolio, index: number) => {
     }
   }
 }
-const removeItem = async (id: number, index: number) => {
+const removeItem = async (id: number) => {
+  if (id < 0) {
+    list.value = list.value.filter((item: IPortfolio) => item.id !== id)
+    list.value = [...list.value]
+    return
+  }
   try {
     const response = await useCustomFetch(`${API_PORTFOLIO}/${id}`, {
       method: 'DELETE',
